User commit

+++ src/main/java/com/example/demo/controller/StudentProfileController.java
@@ -54,7 +54,7 @@ public class StudentProfileController {
     }
 
     @GetMapping("/studentProfile/{id}")
-    public StudentProfile getProfileById(@PathVariable Integer id) {
+    public StudentProfile getProfileById(@PathVariable Long id) {
         return service.getData(id); // exception triggers here
     }
